requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click>=8.1.0,<9.0.0
coverage[toml]==6.4.1
databricks-cli
dbx>=0.7,<0.8
delta-spark==1.1.0
fastapi==0.63.0
flake8
h11==0.12.0
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
mangum==0.10.0
mlflow
packaging==20.9
pandas
pluggy==0.13.1
pre-commit
py==1.10.0
pydantic>=1.0.0,<2.0.0
pyparsing==2.4.7
pyspark==3.2.1
pytest==6.2.4
pytest-cov
pyyaml
requests==2.25.1
scikit-learn
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
